move routes of pages from app.js to /routes/pages.js
